backend/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
python-dotenv==1.2.2
httpx==0.28.1
langchain-openai==1.1.14
langgraph==1.1.8
langgraph-checkpoint-postgres==3.0.5
langsmith==0.7.32
openai==2.32.0
psycopg[binary]==3.3.3
tiktoken==0.12.0
PyJWT==2.12.1
cryptography==46.0.7
SQLAlchemy==2.0.49
alembic==1.18.4
